requirements.txt
albumentations
matplotlib
wxPython
numpy
pillow
opencv
pytorch

